requirements.txt
tensorflow==2.15.0
numpy==1.24.3
flask==3.0.0
flask-cors==4.0.0
pillow==10.1.0
matplotlib==3.8.2
